requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ansi2html==1.8.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
dash==2.11.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.25.1
openai==1.10.0
packaging==23.1
pandas==1.5.1
plotly==5.15.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
websocket-client==1.6.1
Werkzeug==2.2.3
yarl==1.9.2
setuptools~=49.2.0